[AppCheck] Rename reCAPTCHA App Check provider
Renamed `firebase-appcheck-recaptchaenterprise` to
`firebase-appcheck-recaptcha` to standardize the naming
convention. Updated `NetworkClient` to reflect the new `RECAPTCHA`
token type and adjusted internal references accordingly.
