+++ appcheck/firebase-appcheck-recaptcha/existing_api.txt
@@ -1,0 +1,10 @@
+// Signature format: 3.0
+package com.google.firebase.appcheck.recaptcha {
+
+  public class RecaptchaAppCheckProviderFactory implements com.google.firebase.appcheck.AppCheckProviderFactory {
+    method public com.google.firebase.appcheck.AppCheckProvider create(com.google.firebase.FirebaseApp);
+    method public static com.google.firebase.appcheck.recaptcha.RecaptchaAppCheckProviderFactory getInstance(String);
+  }
+
+}
+
