+++ appcheck/firebase-appcheck-recaptcha/firebase-appcheck-recaptcha.gradle
@@ -1,0 +1,69 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+plugins {
+    id 'firebase-library'
+}
+
+firebaseLibrary {
+    libraryGroup = "appcheck-norelease"
+    releaseNotes {
+        name.set("{{app_check}} Recaptcha")
+        versionName.set("appcheck-recaptcha")
+    }
+    previewMode = "norelease"
+}
+
+android {
+    adbOptions {
+        timeOutInMs 60 * 1000
+    }
+
+    namespace "com.google.firebase.appcheck.recaptcha"
+    compileSdkVersion project.compileSdkVersion
+    defaultConfig {
+        targetSdkVersion project.targetSdkVersion
+        minSdkVersion project.minSdkVersion
+        versionName version
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    testOptions.unitTests.includeAndroidResources = false
+}
+
+dependencies {
+    implementation(libs.dagger.dagger)
+
+    api project(':appcheck:firebase-appcheck')
+    api project(':firebase-common')
+    api libs.firebase.components
+    api 'com.google.android.recaptcha:recaptcha:18.7.1'
+
+    annotationProcessor(libs.dagger.compiler)
+
+    testImplementation(project(":integ-testing")) {
+        exclude group: 'com.google.firebase', module: 'firebase-common'
+        exclude group: 'com.google.firebase', module: 'firebase-components'
+    }
+    testImplementation libs.androidx.test.core
+    testImplementation libs.truth
+    testImplementation libs.junit
+    testImplementation libs.mockito.core
+    testImplementation libs.robolectric
+    testImplementation libs.org.json
+}
