+++ appcheck/firebase-appcheck-recaptcha/src/main/java/com/google/firebase/appcheck/recaptcha/internal/ExchangeRecaptchaTokenRequest.java
@@ -1,0 +1,50 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.firebase.appcheck.recaptcha.internal;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+/**
+ * Client-side model of the ExchangeRecaptchaTokenRequest payload from the Firebase App
+ * Check Token Exchange API.
+ */
+public class ExchangeRecaptchaTokenRequest {
+
+  @VisibleForTesting static final String RECAPTCHA_TOKEN_KEY = "recaptchaToken";
+
+  @VisibleForTesting static final String LIMITED_USE_TOKEN_KEY = "limited_use";
+
+  private final String recaptchaToken;
+  private final boolean isLimitedUseToken;
+
+  public ExchangeRecaptchaTokenRequest(@NonNull String recaptchaToken, boolean isLimitedUseToken) {
+    this.recaptchaToken = recaptchaToken;
+    this.isLimitedUseToken = isLimitedUseToken;
+  }
+
+  @NonNull
+  public String toJsonString() throws JSONException {
+    JSONObject jsonObject = new JSONObject();
+    jsonObject.put(RECAPTCHA_TOKEN_KEY, recaptchaToken);
+    if (isLimitedUseToken) {
+      jsonObject.put(LIMITED_USE_TOKEN_KEY, true);
+    }
+
+    return jsonObject.toString();
+  }
+}
