+++ appcheck/firebase-appcheck-recaptcha/src/main/java/com/google/firebase/appcheck/recaptcha/internal/ProviderComponent.java
@@ -1,0 +1,47 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.firebase.appcheck.recaptcha.internal;
+
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.annotations.concurrent.Blocking;
+import com.google.firebase.annotations.concurrent.Lightweight;
+import dagger.BindsInstance;
+import dagger.Component;
+import dagger.Module;
+import java.util.concurrent.Executor;
+import javax.inject.Singleton;
+
+@Singleton
+@Component(modules = ProviderComponent.MainModule.class)
+public interface ProviderComponent {
+  ProviderMultiResourceComponent getMultiResourceComponent();
+
+  @Component.Builder
+  interface Builder {
+    @BindsInstance
+    Builder setFirebaseApp(FirebaseApp firebaseApp);
+
+    @BindsInstance
+    Builder setLiteExecutor(@Lightweight Executor liteExecutor);
+
+    @BindsInstance
+    Builder setBlockingExecutor(@Blocking Executor blockingExecutor);
+
+    ProviderComponent build();
+  }
+
+  @Module
+  abstract class MainModule {}
+}
