+++ subprojects.cfg
@@ -3,7 +3,7 @@ appcheck:firebase-appcheck-debug-testing # sdk
 appcheck:firebase-appcheck-debug # sdk
 appcheck:firebase-appcheck-interop # sdk
 appcheck:firebase-appcheck-playintegrity # sdk
-appcheck:firebase-appcheck-recaptchaenterprise # sdk
+appcheck:firebase-appcheck-recaptcha # sdk
 appcheck:firebase-appcheck # sdk
 
 ai-logic
